fix syn dependency features
`quickcheck_async` uses items from `syn` that are only available if the `full` feature is enabled, however it doesn't specify that in its dependency.

Often that isn't a problem, as many other things depend on `syn` and include that feature, and Cargo's feature unification means `full` gets enabled anyway.  You can reproduce this by removing all the dev-dependencies and running `cargo check`.

Add the feature to the syn dependency explicitly so it's always provided.